Admin services

+++ ConnectUs.Infrastructure/Repositories/MeetupAdminService.cs
@@ -1,9 +1,11 @@
 ﻿using ConnectUs.Domain.Entities;
 using ConnectUs.Domain.IRepositories;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
@@ -20,31 +22,45 @@ namespace ConnectUs.Infrastructure.Repositories
         }
         public async Task CreateAsync(Meetup meetup)
         {
+
             await _context.Meetups.AddAsync(meetup);
             await _context.SaveChangesAsync();
         }
 
-        public async Task Delete(string meetupId)
+        public async Task<bool> Delete(string meetupId, List<Claim> user)
         {
+
             var meetup = _context.Meetups.FirstOrDefault(c => c.Id == Guid.Parse(meetupId));
-            _context.Meetups.Remove(meetup);
-            await _context.SaveChangesAsync();
+            if (meetup.CreatedByUser == user.FirstOrDefault().Value || user[1].Value == "Admin")
+            {
+                _context.Meetups.Remove(meetup);
+                await _context.SaveChangesAsync();
+                return true;
+            }
+            return false;
         }
 
-        public Meetup GetById(string meetupId)
+        public Meetup GetById(string meetupId, List<Claim> user)
         {
             return _context.Meetups.FirstOrDefault(c => c.Id == Guid.Parse(meetupId));
         }
 
-        public IEnumerable<Meetup> GetMeetups(string userId)
+        public IEnumerable<Meetup> GetMeetups(List<Claim> user)
         {
-            return _context.Meetups.ToList().Where(c => c.CreatedByUser == userId);
+            return _context.Meetups.ToList().Where(c => c.CreatedByUser == user.FirstOrDefault().Value);
         }
 
-        public async Task Update(Meetup meetup)
+        public async Task<bool> Update(Meetup meetup, List<Claim> user)
         {
-            _context.Meetups.Update(meetup);
-            await _context.SaveChangesAsync();
+            if (meetup.CreatedByUser == user.FirstOrDefault().Value || user[1].Value == "Admin")
+            {
+
+                _context.Meetups.Update(meetup);
+                await _context.SaveChangesAsync();
+                return true;
+
+            }
+            return false;
 
         }
     }
